% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLab5.R
\name{mybin}
\alias{mybin}
\title{Title}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of success and failure counted}

\item{n}{size of the count}

\item{p}{precentage of success}
}
\value{
vector of the count
}
\description{
Title
}
\examples{
mybin(2, 2, .6) = (1,1) (1,0)
}
