% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFFunction.R
\name{myF}
\alias{myF}
\title{Square function}
\usage{
myF(x)
}
\arguments{
\item{x}{takes vector and return vector squared}
}
\value{
vector of squared components
}
\description{
Square function
}
\examples{
x = 1:30 myF(x)
}
