% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlikf.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Title}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{is used in the outer to define type of variable}

\item{x}{the object that is used for the calculation}

\item{param}{second part of the data}

\item{...}{}
}
\value{
a list of the elements that were calculated by the function
}
\description{
Title
}
