% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLab5.R
\name{mymult}
\alias{mymult}
\title{Title}
\usage{
mymult(iter = 100, n = 10, p = c(1, 1, 1, 1)/4)
}
\arguments{
\item{iter}{number of smaples built}

\item{n}{size of the smaples}

\item{p}{percent of success}
}
\value{
table of maxix success of each test with multimple
varibles included
}
\description{
Title
}
\examples{
mymult(2, 2 , p = c(1,1)/2)
}
