% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{Title}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{is the lower quartile point}

\item{xmax}{maximun density of the graph}

\item{n}{sample size}

\item{alpha}{how large of confidence interval}
}
\value{
a list with the value of p
}
\description{
Title
}
\examples{
mypvalue() will print out a graph with the confidence
interval set at 95 percent degree of freedom of 19 and max density of
4
}
