% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLab5.R
\name{mysample}
\alias{mysample}
\title{Title}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{size of the sample}

\item{iter}{how many time the sample is ran}

\item{time}{on the interval}
}
\value{
vector of the amount of different graphs ran
}
\description{
Title
}
\examples{
sample(5,5,1)
}
